Update editor draft URLs
This commit will update the target URLs for the editor drafts links,
since those URLs were giving HTTP 404. The reason for that is that the
source repository used to build this GitHub Pages site was moved, but
GitHub Pages sites do not redirect when they are transferred to a
different account.